move releated test to InvalidatedIteratorChecker's original testfile
